shared: split out property get helpers
No code changes, just some refactoring.
