+++ src/core/dbus-service.c
@@ -4,7 +4,7 @@
 
 #include "alloc-util.h"
 #include "async.h"
-#include "bus-util.h"
+#include "bus-get-properties.h"
 #include "dbus-cgroup.h"
 #include "dbus-execute.h"
 #include "dbus-kill.h"
