+++ src/import/importd.c
@@ -7,8 +7,8 @@
 
 #include "alloc-util.h"
 #include "bus-common-errors.h"
+#include "bus-get-properties.h"
 #include "bus-log-control-api.h"
-#include "bus-util.h"
 #include "bus-polkit.h"
 #include "def.h"
 #include "fd-util.h"
@@ -22,8 +22,8 @@
 #include "parse-util.h"
 #include "path-util.h"
 #include "process-util.h"
-#include "signal-util.h"
 #include "service-util.h"
+#include "signal-util.h"
 #include "socket-util.h"
 #include "stat-util.h"
 #include "string-table.h"
