+++ src/login/logind-dbus.c
@@ -12,6 +12,7 @@
 #include "bootspec.h"
 #include "bus-common-errors.h"
 #include "bus-error.h"
+#include "bus-get-properties.h"
 #include "bus-locator.h"
 #include "bus-polkit.h"
 #include "bus-unit-util.h"
